requirements.txt
imutils==0.5.2
meson==0.52.0
numpy==1.16.4
pbr==5.4.3
Pillow==6.1.0
pytesseract==0.2.7
six==1.13.0
stevedore==1.31.0
TBB==0.1
virtualenv==16.7.7
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
